Code refactor and README update.
