+++ SmartApp.AV/UI/src/index.html
@@ -100,7 +100,7 @@
     </style>
 
     <div id="content"></div>
-
+    asd
     <script type="text/javascript" src="bundle.js"></script>
 </body>
