+++ SmartApp.AV/UI/src/main.ts
@@ -12,9 +12,9 @@ let factory = new ElfColorfulUIFactory(elem, window);
 
 let kbEventReader = new KBEventReader();
 let readers = [
-	// new TestEventReader(2000),
-	kbEventReader,
-	new TTSEventReader()
+	new TestEventReader(2000),
+	// kbEventReader,
+	// new TTSEventReader()
 ];
 
 let ui = new Builder(factory)
