+++ SmartApp.AV/UI/src/test/testcolorful/TestElfColorfulUI.ts
@@ -5,7 +5,8 @@ import { ElfUI, ElfUIFactory } from '../../ui/ElfUI';
 import { IEmotion, Emotion } from '../../emotion/Emotion';
 import { UIWidget, UIWidgetFactory, EmotionalWidget } from '../../ui/widget/UIWidget';
 import { TestFace } from './face/TestFace';
-import { IContent, ContentFactory, DefaultContentFactory, AudioContent } from '../../content/Content';
+import { IContent, ContentFactory, AudioContent } from '../../content/Content';
+import { DefaultContentFactory } from '../../content/DefaultContentFactory';
 import { DefaultUIWidget } from './widget/DefaultTestUIWidget';
 import { TextTestUIWidget } from './widget/TextTestUIWidget';
 import { TestColorfulUIWidgetFactory } from './TestColorfulUIWidgetFactory';
